+++ HISTORY.md
@@ -10,6 +10,7 @@
 
 ### Fixed
 * Prune workers which haven't been registered but have set a heartbeat
+* `Resque::Failure::Multiple.remove` did not pass on the queue parameter
 
 ## 2.0.0 (2018-11-06)